+++ packages/ui-components/src/lib/components/charts/TableTimeFilters.test.svelte
@@ -1,0 +1,9 @@
+<script>
+	import TableTimeFilters from './TableTimeFilters.svelte';
+	import { writable } from 'svelte/store';
+
+	export let startTimeStore = writable();
+	export let endTimeStore = writable();
+</script>
+
+<TableTimeFilters bind:startTimestamp={$startTimeStore} bind:endTimestamp={$endTimeStore} />
