--- tauri-app/src/lib/components/charts/TableTimeFilters.test.svelte
@@ -1,9 +1,0 @@
-<script>
-  import TableTimeFilters from './TableTimeFilters.svelte';
-  import { writable } from 'svelte/store';
-
-  export let startTimeStore = writable();
-  export let endTimeStore = writable();
-</script>
-
-<TableTimeFilters bind:startTimestamp={$startTimeStore} bind:endTimestamp={$endTimeStore} />
